refactor(scheduler): reduce abstract api surface
